jwt + api correct api calls

+++ project.api/Core/Services/SkillServices.cs
@@ -1,4 +1,5 @@
-﻿using project.api.Core.Interfaces;
+﻿using Microsoft.AspNetCore.Http;
+using project.api.Core.Interfaces;
 using project.api.Data.Context;
 using project.api.Data.Models;
 using System;
@@ -11,29 +12,55 @@ namespace project.api.Core.Services
     public class SkillServices : ISkillServices
     {
         private readonly AppDBContext _context;
-        public SkillServices(AppDBContext context)
+        private readonly User _user;
+        public SkillServices(AppDBContext context, IHttpContextAccessor httpContextAccessor)
         {
             _context = context;
+            _user = _context.Users
+                       .First(u => u.Username == httpContextAccessor.HttpContext.User.Identity.Name);
         }
 
         public Skill CreateSkill(Skill skill)
         {
-            _context.Add(skill);
-            _context.SaveChanges();
-
+            var app = _context.Applications.Where(a => a.ApplicationId == skill.ApplicationId).First();
+            if (app.CandidateId == _user.Id)
+            {
+                _context.Add(skill);
+                _context.SaveChanges();
+               // return skill;
+            }
             return skill;
+            //return skill;
         }
 
-        public void DeleteSkill(Skill skill)
+        public bool DeleteSkill(Skill skill)
         {
-            _context.Skills.Remove(skill);
-            _context.SaveChanges();
+            var dbSkill = _context.Skills.Where(s => s.Application.CandidateId == _user.Id && s.SkillId == skill.SkillId).FirstOrDefault();
+            if (dbSkill != null)
+            {
+                _context.Skills.Remove(skill);
+                _context.SaveChanges();
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
 
-        public void EditSkill(Skill skill)
+        public bool EditSkill(Skill skill)
         {
-            _context.Skills.Update(skill);
+            var dbSkill = _context.Skills.Where(s => s.Application.CandidateId == _user.Id && s.SkillId == skill.SkillId).FirstOrDefault();
+            if (dbSkill != null)
+            {
+                _context.Skills.Update(skill);
             _context.SaveChanges();
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
 
         public Skill GetSkill(int id)
@@ -44,7 +71,7 @@ namespace project.api.Core.Services
 
         public List<Skill> GetSkills()
         {
-            return _context.Skills.ToList();
+            return _context.Skills.Where(s => s.Application.CandidateId == _user.Id).ToList();
         }
 
         public List<Skill> GetSkillsFromApplication(int applicationId)
